+++ src/attr/_make.py
@@ -368,6 +368,8 @@ def attributes(maybe_cls=None, these=None, repr_ns=None,
                 cls = _add_pickle(cls)
         if slots is True:
             cls_dict = dict(cls.__dict__)
+            was_unhashable = ('__hash__' in cls_dict and
+                              cls_dict['__hash__'] is None)
             cls_dict["__slots__"] = tuple(ca_list)
             for ca_name in ca_list:
                 # It might not actually be in there, e.g. if using 'these'.
@@ -376,6 +378,14 @@ def attributes(maybe_cls=None, these=None, repr_ns=None,
 
             qualname = getattr(cls, "__qualname__", None)
             cls = type(cls)(cls.__name__, cls.__bases__, cls_dict)
+
+            if (not was_unhashable and hash is False and
+                    '__hash__' in cls.__dict__ and
+                    cls.__dict__['__hash__'] is None):
+
+                # Python just set __hash__ = None for us, revert this.
+                del cls.__hash__
+
             if qualname is not None:
                 cls.__qualname__ = qualname